pyproject.toml
[project]
name = "pubg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pyinstaller>=6.17.0",
    "pynput>=1.8.1",
    "pystray>=0.19.5",
]
